Fixing errors and neatening code.
Re-worded several functions so that rasterOut is the variable returned.
This wording should be consistent throughout.

Changed break condition for k-means and fuzzy c-means from "<" to "<="
as this allows the break condition to be set to 0 (no change between
iterations) while still allowing the algorithm to break early.

Principal component can now read pre-defined eigenvectors. This is
fairly limited in usage, but  it could be useful to force the same
transformation across several images/subsets of a larger image.