Initial refactoring of focal functions.
Initial version of the FocalCalc function, which is a refactoring of the
existing focal commands - in addition it works for multi-layered
rasters. This should replace most of the current FocalSummary,
FocalAngle, GaussianSmooth and WMat.

Additionally starting to look at modifying the temporary file pattern
used. May be a good candidate for replacing with a function, to make
sure they're all the same pattern.

Working on a few updates for other functions, once they're ready several
older functions will be pruned/updated.