Removed export functions. Re-configured temp files.
Removed ExportKML and ExportRGB. Neither produced particularly useful
products - in the case of ExportKML, there is little to no added
functionality over the raster KML function. ExportRGB can be handled
much better in other open-source software, such as QGIS. If either is
re-created, it will have to be with significantly more functionality.

Updated the way temporary files are handled, is now wrapped in a
function, so any future changes will automatically apply to other
functions.