k & c means updates - writing at end only
Changed k-means and fuzzy c-means functions to only write to disk on the
last iteration (or once the loop has converged. Results in dropping
approximately a quarter of the calculation time (depending on how many
iterations were required). The handling of centroids is one potential
issue that will need to be reviewed to make sure they are calculated on
each iteration (and only once!).