Updates to LS processing, Reverted RasterLoad
Reworked LandsatProcessing - better detection of sensors, and fixed a
few specific crashes. Potentially sped up the central part of the
masking step as well. Not quite where I want to end up her as I'd rather
the output specified which file it came from (to differentiate files in
browsers such as QGIS) - though this isn't a major concern.

Also experimenting with a slightly different method for outputting
progress, three dots at the end of the line represent loading data,
processing and writing result to disk. Not overly useful, but adds an
extra layer of information without overloading things.

Also reverted changes to RasterLoad - while exporting as  a specific
format made a certain amount of sense, the consistency of always
returning a RasterStack is probably better.