OIF and improved progress reporting.
Added optimal index factor - mainly as a visualisation aid.

Modified how progress reporting is handled too. Adds dots after the
block number as it progresses, the first after data is loaded, the
second after it is processed and the third after it is written to disk
(if it is). Adds extra granularity to returned information.