Cleaned up style to-do's
Made an attempt to fix all the style issues I had marked as needing
attention. Generally happy with the results, except for FocalCalc -
though if/when I re-write it with the parallel library, it should be
cleared up somewhat by that anyway.

Geomorphometry still needs some work, but the same applies to that as
FocalCalc.